refactor: create a RaceCars class
- move all created functions to the RaceCars class and leave main just as start point
- separte input for car names and rounds from main into their own functions
- code cleanup following Kotlin coding conventions

The creation of a different class for races incluggs the Application class and makes it easier to navigate
